+++ Plugins/Library/src/main/java/com/tagnumelite/projecteintegration/plugins/library/PluginRebornCore.java
@@ -21,20 +21,19 @@
  */
 package com.tagnumelite.projecteintegration.plugins.library;
 
+import com.tagnumelite.projecteintegration.api.internal.lists.InputList;
 import com.tagnumelite.projecteintegration.api.mappers.PEIMapper;
 import com.tagnumelite.projecteintegration.api.plugin.APEIPlugin;
 import com.tagnumelite.projecteintegration.api.plugin.PEIPlugin;
 import com.tagnumelite.projecteintegration.api.utils.ConfigHelper;
-import moze_intel.projecte.emc.IngredientMap;
-import net.minecraft.item.ItemStack;
 import reborncore.api.recipe.IBaseRecipeType;
 import reborncore.api.recipe.RecipeHandler;
-import reborncore.common.recipes.RecipeTranslator;
+import reborncore.common.recipes.IRecipeInput;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 @PEIPlugin("reborncore")
 public class PluginRebornCore extends APEIPlugin {
@@ -44,52 +43,37 @@ public class PluginRebornCore extends APEIPlugin {
     }
 
     private class RebornCoreMapper extends PEIMapper {
-        private final Map<String, Boolean> RECIPE_CFG_MAP = new HashMap<>();
-
         public RebornCoreMapper() {
             super("Reborn Core");
         }
 
         @Override
         public void setup() {
-            for (IBaseRecipeType recipe : RecipeHandler.recipeList) {
-                String recipe_name = recipe.getRecipeName();
-                if (RECIPE_CFG_MAP.containsKey(recipe_name)) {
-                    if (!RECIPE_CFG_MAP.get(recipe_name))
-                        continue;
-                } else {
-                    RECIPE_CFG_MAP.put(recipe_name, config.getBoolean(ConfigHelper.getConversionName(recipe_name),
-                        category, true, "Enable conversions for machine " + recipe_name));
+            // Convert List<IBaseRecipeType> into map of IBaseRecipeType.getRecipeName and List<IBaseRecipeType>
+            Map<String, List<IBaseRecipeType>> recipe_map = RecipeHandler.recipeList.stream()
+                .collect(Collectors.groupingBy(IBaseRecipeType::getRecipeName));
+
+            for (String name : recipe_map.keySet()) {
+                if (config.getBoolean(ConfigHelper.getConversionName(name), category,
+                    true, "Enable conversions for machine " + name)) {
+                    for (IBaseRecipeType recipe : recipe_map.get(name)) {
+                        List<Object> outputs = new ArrayList<>(recipe.getOutputs());
+                        if (outputs.isEmpty())
+                            continue;
+
+                        List<Object> inputs = new ArrayList<>();
+                        for (Object input : recipe.getInputs()) {
+                            if (input instanceof IRecipeInput) {
+                                inputs.add(new InputList<Object>(((IRecipeInput) input).getAllStacks()));
+                            } else {
+                                inputs.add(input);
+                            }
+                        }
+
+                        addRecipe(outputs, inputs);
+                    }
                 }
-
-                List<Object> outputs = new ArrayList<>(recipe.getOutputs());
-                if (outputs == null || outputs.isEmpty())
-                    continue;
-
-                List<Object> inputs = recipe.getInputs();
-                if (inputs == null || inputs.isEmpty())
-                    continue;
-
-                IngredientMap<Object> ingredients = new IngredientMap<Object>();
-
-                inputs.forEach(input -> {
-                    ItemStack item = RecipeTranslator.getStackFromObject(input);
-                    ingredients.addIngredient(item, item.getCount());
-                });
-
-                Map<Object, Integer> map = ingredients.getMap();
-                if (map == null || map.isEmpty())
-                    continue;
-
-                addConversion(outputs, map);
-				/*
-				outputs.forEach(output -> {
-					if (output != null && !output.isEmpty())
-						addConversion(output, map);
-				});*/
             }
-
-            RECIPE_CFG_MAP.clear();
         }
     }
 }