+++ Plugins/Library/src/main/java/com/tagnumelite/projecteintegration/plugins/library/PluginSonarCore.java
@@ -23,7 +23,6 @@ package com.tagnumelite.projecteintegration.plugins.library;
 
 import com.tagnumelite.projecteintegration.api.mappers.PEIMapper;
 import sonar.core.recipes.DefaultSonarRecipe;
-import sonar.core.recipes.ISonarRecipeObject;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -34,31 +33,14 @@ public class PluginSonarCore {
             super(name);
         }
 
-        public SonarMapper(String name, String description) {
-            super(name, description);
-        }
-
         protected void addRecipe(DefaultSonarRecipe recipe) {
             List<Object> inputs = new ArrayList<>();
-
-            for (ISonarRecipeObject input : recipe.inputs()) {
-                if (input.getJEIValue().size() == 1)
-                    inputs.add(input.getJEIValue().get(0));
-                else
-                    inputs.add(input.getJEIValue());
-            }
+            recipe.inputs().forEach(i -> inputs.add(i.getJEIValue()));
 
             ArrayList<Object> outputs = new ArrayList<>();
+            recipe.outputs().forEach(o -> outputs.addAll(o.getJEIValue()));
 
-            for (ISonarRecipeObject out : recipe.outputs()) {
-                outputs.addAll(out.getJEIValue());
-				/*for (ItemStack output : out.getJEIValue()) {
-					addRecipe(output, inputs.stream().toArray());
-
-				}*/
-            }
-
-            addRecipe(outputs, inputs.toArray());
+            addRecipe(outputs, inputs);
         }
     }
 }